+++ src/utils/index.js
@@ -1,5 +1,5 @@
 export { uuid } from './uuid';
 export { linearEase } from './linearEase';
 export { getPath } from './getPath';
-export { getWrapperStyle, getTimeStyle, svgStyle } from './styles';
+export { getWrapperStyle, getTimeStyle, svgStyle, visuallyHidden } from './styles';
 export { getColorsRGB, getNormalizedColors, getStroke, colorsValidator } from './colors';