+++ src/utils/styles.js
@@ -26,3 +26,16 @@ export const svgStyle = {
   width: '100%',
   height: '100%'
 };
+
+export const visuallyHidden = {
+  position: 'absolute',
+  width: '1px',
+  height: '1px',
+  margin: '-1px',
+  border: 0,
+  clip: 'rect(0 0 0 0)',
+  padding: 0,
+  overflow: 'hidden',
+  whiteSpace: 'nowrap',
+  wordWrap: 'normal'
+};